ESLint mechanical fixes.
Extracted from https://github.com/microsoft/vcpkg-tool/pull/690

* Deleted a bunch of unreferenced files.
* Fixed files that were CRLF->LF.
* Fixed a few spelling mistakes.
