--- ce/ce/interfaces/metadata/registries/git-registry.ts
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-import { ArtifactRegistry } from './artifact-registry';
-
-
-export interface GitRegistry extends ArtifactRegistry {
-}
