--- ce/ce/registries/https-registry.ts
@@ -1,3 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
