+++ ce/ce/session.ts
@@ -1,4 +1,3 @@
-/* eslint-disable prefer-const */
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
@@ -401,5 +400,4 @@ export class Session {
     }
     return value;
   }
-
 }
