+++ ce/ce/util/uri.ts
@@ -340,4 +340,3 @@ export function isFilePath(uriOrPath?: Uri | string): boolean {
   }
   return false;
 }
-
